+++ src/main/java/com/hmh/mmp/dto/cash/CashSaveDTO.java
@@ -17,7 +17,7 @@ public class CashSaveDTO {
     private String cashName;
     private Long memberId;
     private String cashMemo;
-    private Long totalAsset;
+    private double totalAsset;
     private String cashPhotoName;
     private MultipartFile cashPhoto;
 }
